Multiplayer + screen changes
I added multipler (with 1 player having wasd and the other have up down left right arrow keys). I made the players have a color (red and green) and the ladder is grey. Also moved the ladder's y positon to be more inline with the camera, and i changed the friction on the ladder as it kept moving much further for me than it should have.