Added Ladder Sprite But Bugs in Code!
I added a free Ladder Sprite (made 2 of them in the scene), I commented out the print statements in the trigger script. However, sometimes when a player is grabbing onto the ladder above ground height they can still move the ladder and they can get off the ladder and fly.